# ArzTiServer/wwwroot/swagger-original.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Webservice ArzTI API",
    "description": "Kommunikation des Webservers mit den jeweiligen Rechenzentren ",
    "version": "V1"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Local server"
    },
    {
      "url": "http://arzsoftware.de/v1",
      "description": "Main (production) server"
    },
    {
      "url": "http://staging-api.arzsoftware.de",
      "description": "Internal staging server for testing"
    }
  ],
  "paths": {
...
    "/rezept/{ruid}": {
      "get": {
        "tags": [
          "RezeptUID"
        ],
        "summary": "liefert ein Rezept",
        "description": "",
        "operationId": "getRezeptUId",
        "parameters": [
          {
            "name": "ruid",
            "in": "path",
            "required": true,
            "description": "Die Rezept UID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert das gefundene Rezepte.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "RezeptUID"
        ],
        "summary": "löscht ein Rezept",
        "description": "",
        "operationId": "deleteRezeptUId",
        "parameters": [
          {
            "name": "ruid",
            "in": "path",
            "required": true,
            "description": "Die Rezept UID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert das Rezept für die ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Rezept"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_write"
            ]
          }
        ]
      }
    },
    "/rezept/{ruid}/status": {
      "get": {
        "tags": [
          "RezeptUID"
        ],
        "summary": "liefert den Status eines Rezepts",
        "description": "",
        "operationId": "getRezeptUIdStatus",
        "parameters": [
          {
            "name": "ruid",
            "in": "path",
            "required": true,
            "description": "Die Rezept UID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert das gefundene Rezepte.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "RezeptUID"
        ],
        "summary": "Statusänderung eines Rezeptes",
        "description": "Statusänderung eines Rezeptes",
        "operationId": "patchRezeptUIdStatus",
        "parameters": [
          {
            "name": "ruid",
            "in": "path",
            "required": true,
            "description": "Die Rezept UID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "description": "Status value",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liefert den Status des Rezeptes für die ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "RezeptStatus": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "status": {
            "type": "string",
            "description": "Status des Rezepts"
          }
        }
      },
      "RezeptTyp": {
        "type": "string",
        "enum": [
          "ERezept",
          "MRezept",
          "PRezept"
        ]
      },
      "Rezept": {
        "title": "Rezept",
        "type": "object",
        "description": "base64 codierte Leistung",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "RezeptId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept ID",
        "properties": {
          "typ": {
            "$ref": "#/components/schemas/RezeptTyp"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "RezeptUId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept RUID",
        "properties": {
          "ruid": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}